+++ src/cell.js
@@ -47,6 +47,8 @@ export class Cell {
         this.w.activeP = this
         this.e.style.borderColor = FOCUSED_BORDER_COLOR
         this.w.toggleDivideButtons()
+        setTimeout(() => window.location.href = `#pane-${this.id}`)
+        this.w.nameE.setAttribute("href", `#pane-${this.id}`)
     }
     /*
      * Used to grow/shrink the terminal based on containing element dimensions
@@ -137,17 +139,25 @@ export class Cell {
             canvas.height = 0;
             canvas.width = 0;
         })
-
+        if (this.zoomed)
+            this.toggleZoom()
         this.e.remove()
         if (this.layout)
             this.layout.onClose(this)
         this.gate.sendState()
     }
     styleZoomed(e) {
-        let H = document.body.offsetHeight
-        e.style.height = `${H - 42}px`
+        e = e || this.t7.zoomedE.querySelector(".pane")
+        const se = this.gate.e.querySelector(".search-box")
+        let style
+        if (se.classList.contains("show"))
+            style = `${document.querySelector('.windows-container').offsetHeight - 22}px`
+        else
+            style = `${document.body.offsetHeight - 42}px`
+        e.style.height = style
         e.style.top = "0px"
         e.style.width = "100%"
+        this.fit()
     }
     toggleZoom() {
         if (this.zoomed) {
@@ -170,12 +180,11 @@ export class Cell {
             e.style.borderColor = FOCUSED_BORDER_COLOR
             e.appendChild(te)
             c.appendChild(e)
-            this.styleZoomed(e)
             this.catchFingers(e)
             document.body.appendChild(c)
             this.t7.zoomedE = c
             this.w.e.classList.add("hidden")
-            this.resizeObserver = new window.ResizeObserver(_ => this.styleZoomed(e))
+            this.resizeObserver = new window.ResizeObserver(() => this.styleZoomed(e))
             this.resizeObserver.observe(e);
         }
         this.zoomed = !this.zoomed