+++ src/terminal7.js
@@ -223,7 +223,7 @@ echo "${fp}" >> ~/.config/webexec/authorized_fingerprints`
 
 
         document.addEventListener("keydown", ev => {
-            if (ev.key == "Meta") {
+            if ((ev.key == "Meta") && (Capacitor.getPlatform() != "ios")) {
                 this.metaPressStart = Date.now()
                 this.run(() => this.showKeyHelp(), terminal7.conf.ui.quickest_press)
             } else
@@ -570,6 +570,7 @@ peer_name = "${peername}"\n`
         document.querySelectorAll(".pane-buttons").forEach(
             e => e.classList.add("off"))
         window.location.href = "#home"
+        document.title = "Terminal 7"
     }
     /* 
      * Terminal7.logDisplay display or hides the notifications.
@@ -957,7 +958,6 @@ peer_name = "${peername}"\n`
             g.online = m.peer_update.online
             return
         }
-        // next to are for connected session, ignore if no session
         if (!g.session) {
             console.log("session is close ignoring message", m)
             return
@@ -1005,7 +1005,10 @@ peer_name = "${peername}"\n`
         this.firstPointer = null
         this.lastT = null
         this.gesture = null
-        this.longPressGate = null
+        if (this.longPressGate) {
+            clearTimeout(this.longPressGate)
+            this.longPressGate = null
+        }
         return
     }
     onPointerDown(ev) {